requirements.txt
djangorestframework==3.6.3
django-markdown-deux==2.6.8
django-pagedown==0.1.3
markdown2==2.3.1
Pillow==4.2.0